[REFACTOR] Layer First to Feature First in Unit Test
