+++ test/features/auth/data/repositories/auth_repository_impl_test.dart
@@ -1,0 +1,100 @@
+import 'dart:convert';
+
+import 'package:dartz/dartz.dart';
+import 'package:flutter_auth_app/core/core.dart';
+import 'package:flutter_auth_app/dependencies_injection.dart';
+import 'package:flutter_auth_app/features/features.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:mockito/mockito.dart';
+
+import '../../../../helpers/json_reader.dart';
+import '../../../../helpers/paths.dart';
+import '../../../../helpers/test_mock.mocks.dart';
+
+void main() {
+  late MockAuthRemoteDatasource mockAuthRemoteDatasource;
+  late AuthRepositoryImpl authRepositoryImpl;
+  late Login login;
+  late Register register;
+
+  setUp(() async {
+    await serviceLocator(isUnitTest: true);
+    mockAuthRemoteDatasource = MockAuthRemoteDatasource();
+    authRepositoryImpl = AuthRepositoryImpl(mockAuthRemoteDatasource);
+    login = LoginResponse.fromJson(json.decode(jsonReader(successLoginPath)))
+        .toEntity();
+    register =
+        RegisterResponse.fromJson(json.decode(jsonReader(successRegisterPath)))
+            .toEntity();
+  });
+
+  group("login", () {
+    final loginParams = LoginParams(email: "email", password: "password");
+    test('should return login when call data is successful', () async {
+      // arrange
+      when(mockAuthRemoteDatasource.login(loginParams)).thenAnswer(
+        (_) async => LoginResponse.fromJson(
+          json.decode(jsonReader(successLoginPath)),
+        ),
+      );
+
+      // act
+      final result = await authRepositoryImpl.login(loginParams);
+
+      // assert
+      verify(mockAuthRemoteDatasource.login(loginParams));
+      expect(result, equals(Right(login)));
+    });
+
+    test(
+      'should return server failure when call data is unsuccessful',
+      () async {
+        // arrange
+        when(mockAuthRemoteDatasource.login(loginParams))
+            .thenThrow(ServerException(''));
+
+        // act
+        final result = await authRepositoryImpl.login(loginParams);
+
+        // assert
+        verify(mockAuthRemoteDatasource.login(loginParams));
+        expect(result, equals(const Left(ServerFailure(''))));
+      },
+    );
+  });
+
+  group("register", () {
+    final registerParams = RegisterParams(email: "email", password: "password");
+    test('should return register when call data is successful', () async {
+      // arrange
+      when(mockAuthRemoteDatasource.register(registerParams)).thenAnswer(
+        (_) async => RegisterResponse.fromJson(
+          json.decode(jsonReader(successRegisterPath)),
+        ),
+      );
+
+      // act
+      final result = await authRepositoryImpl.register(registerParams);
+
+      // assert
+      verify(mockAuthRemoteDatasource.register(registerParams));
+      expect(result, equals(Right(register)));
+    });
+
+    test(
+      'should return server failure when call data is unsuccessful',
+      () async {
+        // arrange
+        when(mockAuthRemoteDatasource.register(registerParams))
+            .thenThrow(ServerException(''));
+
+        // act
+        final result = await authRepositoryImpl.register(registerParams);
+
+        // assert
+        verify(mockAuthRemoteDatasource.register(registerParams));
+        expect(result, equals(const Left(ServerFailure(''))));
+      },
+    );
+  });
+}
