+++ test/features/users/presentation/dashboard/cubit/users_cubit_test.mocks.dart
@@ -1,0 +1,59 @@
+// Mocks generated by Mockito 5.3.2 from annotations
+// in flutter_auth_app/test/features/users/presentation/dashboard/cubit/users_cubit_test.dart.
+// Do not manually edit this file.
+
+// ignore_for_file: no_leading_underscores_for_library_prefixes
+import 'dart:async' as _i4;
+
+import 'package:dartz/dartz.dart' as _i2;
+import 'package:flutter_auth_app/core/core.dart' as _i5;
+import 'package:flutter_auth_app/features/users/users.dart' as _i3;
+import 'package:mockito/mockito.dart' as _i1;
+
+// ignore_for_file: type=lint
+// ignore_for_file: avoid_redundant_argument_values
+// ignore_for_file: avoid_setters_without_getters
+// ignore_for_file: comment_references
+// ignore_for_file: implementation_imports
+// ignore_for_file: invalid_use_of_visible_for_testing_member
+// ignore_for_file: prefer_const_constructors
+// ignore_for_file: unnecessary_parenthesis
+// ignore_for_file: camel_case_types
+// ignore_for_file: subtype_of_sealed_class
+
+class _FakeEither_0<L, R> extends _i1.SmartFake implements _i2.Either<L, R> {
+  _FakeEither_0(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
+
+/// A class which mocks [GetUsers].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockGetUsers extends _i1.Mock implements _i3.GetUsers {
+  MockGetUsers() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  _i4.Future<_i2.Either<_i5.Failure, _i3.Users>> call(
+          _i3.UsersParams? params) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #call,
+          [params],
+        ),
+        returnValue: _i4.Future<_i2.Either<_i5.Failure, _i3.Users>>.value(
+            _FakeEither_0<_i5.Failure, _i3.Users>(
+          this,
+          Invocation.method(
+            #call,
+            [params],
+          ),
+        )),
+      ) as _i4.Future<_i2.Either<_i5.Failure, _i3.Users>>);
+}
