+++ test/helpers/test_mock.dart
@@ -1,9 +1,10 @@
-import 'package:flutter_auth_app/data/datasources/datasources.dart';
-import 'package:flutter_auth_app/domain/domain.dart';
+import 'package:flutter_auth_app/features/features.dart';
 import 'package:mockito/annotations.dart';
 
 @GenerateMocks([
   AuthRepository,
   AuthRemoteDatasource,
+  UsersRepository,
+  UsersRemoteDatasource,
 ])
 void main() {}
