--- test/presentation/page/auth/cubit/auth_cubit_test.mocks.dart
@@ -1,39 +1,0 @@
-// Mocks generated by Mockito 5.2.0 from annotations
-// in flutter_auth_app/test/presentation/page/auth/cubit/auth_cubit_test.dart.
-// Do not manually edit this file.
-
-import 'dart:async' as _i4;
-
-import 'package:dartz/dartz.dart' as _i2;
-import 'package:flutter_auth_app/core/core.dart' as _i5;
-import 'package:flutter_auth_app/domain/domain.dart' as _i3;
-import 'package:mockito/mockito.dart' as _i1;
-
-// ignore_for_file: type=lint
-// ignore_for_file: avoid_redundant_argument_values
-// ignore_for_file: avoid_setters_without_getters
-// ignore_for_file: comment_references
-// ignore_for_file: implementation_imports
-// ignore_for_file: invalid_use_of_visible_for_testing_member
-// ignore_for_file: prefer_const_constructors
-// ignore_for_file: unnecessary_parenthesis
-// ignore_for_file: camel_case_types
-
-class _FakeEither_0<L, R> extends _i1.Fake implements _i2.Either<L, R> {}
-
-/// A class which mocks [PostLogin].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockPostLogin extends _i1.Mock implements _i3.PostLogin {
-  MockPostLogin() {
-    _i1.throwOnMissingStub(this);
-  }
-
-  @override
-  _i4.Future<_i2.Either<_i5.Failure, _i3.Login>> call(
-          _i3.LoginParams? params) =>
-      (super.noSuchMethod(Invocation.method(#call, [params]),
-              returnValue: Future<_i2.Either<_i5.Failure, _i3.Login>>.value(
-                  _FakeEither_0<_i5.Failure, _i3.Login>()))
-          as _i4.Future<_i2.Either<_i5.Failure, _i3.Login>>);
-}
